从EasyX移植到SDL2，从sln移植到CMake 添加web版本支持 使用sdl2硬件加速，提升性能表现 优化平台碰撞逻辑，确保不会出现在帧率不稳定时出现角色无视平台下落的问题 创建了MessageBox和Button类，不调用外部的提示框 将yellowmashroom改名为buttermashroom Implement Player class with movement, animations, and input handling; add main game loop and resource loading in main.cpp; configure vcpkg dependencies for JSON and SDL libraries.

+++ head/bullets/Bubble.h
@@ -1,8 +1,8 @@
 #pragma once
 
 #include "Bullet.h"
-#include "../Camera.h"
-#include "../players/Player.h"
+#include "baseItem/Camera.h"
+#include "players/Player.h"
 
 extern Atlas atlas_bubble;
 
@@ -23,7 +23,7 @@ public:
         });
     }
     ~Bubble()=default;
-    void update(int delta)override{
+    void update(float delta)override{
         animation_bubble.update(delta);
         if(isValid){
             Vector2 newPos=owner->getPosition();
